rename auth module to authentication
