+++ apps/server/src/authentication/dto/user.dto.ts
