+++ apps/server/src/authentication/entities/auth.entity.ts
