+++ apps/server/src/authentication/guards/jwt-auth.guard.ts
