+++ apps/server/src/authentication/interfaces/jwt-payload.ts
