+++ apps/server/src/authentication/jwt-options.provider.ts
