+++ apps/server/src/authentication/strategies/jwt.strategy.ts
